Allow SpringMotorDefinitionParser to create nested beans
Instead of creating beans registered in the top application context, it
is sometimes useful to create nested beans. This is done by returning
the created bean definition from the parse method instead of registering
the bean with the context registry.

The behaviour is determined by the tag used to define the motor

    motor -> existing behaviour, bean is registered
    imotor -> return bean definition but do not register anything
        (inner motor)

This allows configuration similar to the following

    <bean id="motors" class="...ScannableGroup" >
        <property name="groupMembers">
            <list>
	        <gda:imotor id="x" live-pvName="BLXXI-MO-X" />
	        <gda:imotor id="y" live-pvName="BLXXI-MO-Y" />
	    </list>
	</property>
    </bean>

Change-Id: I1f24d3e72763b2b05c97c9042009b03387169722