Prevent invalid client registration with BatonManager
If a beamline is using UIScanDataPointEventService it attempted to
dispose the instance as part of bundle shutdown. For the case where the
client hasn't started properly (due to e.g. no visits found) this caused
the client to be registered in the BatonManager after it had closed.

Also make initialisation method synchronized.

Change-Id: If4e2733a6ee49392cd6cb84bfc794d7b1b89507b