K11: Fix radio button selection
The existing logic would ensure the first option is selected, which is
correct for the first time the code is executed. Subsequent times, such
as when the perspective is switched and switched back, we need to honour
the existing selection.

Additionally fixes SWTExceptions by returning early when widgets are
disposed.

Change-Id: I9ed00c4ae87cb87165d550e8cb6023cbc6ed4f9c