[DAQ-3717] add submitScan() function to Jython template for copy scan
The existing template was sufficient for the use case where the user
wished to copy the scan command into a script. However, many users may
want to simply run the scan command, and this was not easy to do.

A new function, 'submitScan' has now been added that will submit and
run the given scan.

Change-Id: Ief1f3381274116150840e267098cabb6cd33f470