K11: Redefine AcquisitionKeys and a 'beam selector' scan
In order for the existing perspectives to manage saving and loading of
beam selector scans, the components that save/load/display saved scans
want these scans to be of type AcquisitionPropertyType.DIFFRACTION,
because the client shows beam selector scans as a type of diffraction
measurements.

For that reason, AcquisitionKeys gets a new AcquisitionSubType field and
BEAM_SELECTOR gets moved from AcquisitionPropertyType to this new type.

The problem that immediately follows this change is that scan properties
are currently defined at the AcquisitionPropertyType level, and beam
selector scans and diffraction scans have different properties. Instead
of defining a new property class for 'subtypes', this change introduces
AcquisitionManager which holds AcquisitionTemplates for different scans.
These templates are defined in Spring beans and set in the manager which
is declared as an Osgi component. The idea is to gradually migrate from
a properties-based scan definition to this template system which seems
more consistent with the rest of GDA.

DocumentFactory now gets a new method to create a new acquisition from
the aforementioned templates. This will eventually replace completely
the method it overloads.

Change-Id: Ibb4054a2cfa3731a7aa4df85d7605264a57d6276