Ensure that Jython exceptions without tracebacks are printed
When Jython itself fails (rather than the code it's running), the
exception that is thrown doesn't always have a traceback accessible
from the PyException that wraps it. In these cases, the exception
wouldn't be printed and users would see no output from a script/command
that failed.

An example of this is running a script that compiles to a function
larger than Java's 64kB limit. See DAQ-3628.

Change-Id: I70af79e0254c9425a4cb9b0bcf9c96f04198400e