Only show GdaBuiltin usage if args are not valid
Instead of assuming all TypeErrors raised when calling a builtin
function are from invalid usage, check if the provided arguments match
any of the provided functions.

If the arguments are valid, the original exception is rethrown as the
wrapped function should have been called, otherwise show the usage
information as before.

This adds a lot of overhead and duplicates a lot of what is done already
by Jython (unfortunately in private classes) but is only called in the
error case where a useful error message is more important than
performance.

See DAQ-3415 and DAQ-3809
Original bug caused by fix for DAQ-2943

Change-Id: I48ef6e9213714c87c1a2d4edb351930132bc866b