DAQ-3840 : Determine names of available Xspress PVs dynamically
Different IOC versions of Xspress3 have undergone several different
naming conventions for retrieving scaler array data. This change
dynamically determines which PV names to use by examining which PVs
are actually available.

- The PV name and start/stop/update values for different Time series
array versions are now stored in an Enum. The Enum to use is determined
during server startup by looping over the Enum examining the the
available PVs using channel access.
- start/stop/update timeseries methods have been added to the EpicsPV
provider, so that the Xspress3Controller logic is simplified.
- The pvUpdate variable uses ':UPDATE' or ':UPDATE_ARRAYS' suffix
depending on which PV is present. A DeviceException is thrown if neither
is present.
- The 'useNewEpicsInterface' flag has been removed, since it is now
redundant.

Change-Id: I0e58f07b3ae37a035d4c73cad91f8657196a972c