K11: Instantiate responder servlet manually
At some point this became a Spring Component but it was found that
Spring instantiated two instances (I do not know why) which leads to
two NodeFileCreator handlers per request. This will now be instantiated
explicitly through Spring beans.

Since it is no longer an autowired component, references to it are
removed in Spring test configuration filters.

The responder is here also renamed to NodeFileCreationgRequestResponder.
It should have been renamed when ExperimentStructureJobRequest was
renamed to NodeFileCreationRequest.

Change-Id: I9b37cd1870de31bcb7be01f4c953d5dccbcafa34