I15-680 (9.23) Replace ScriptChooserView with an unconfigured CommissioningView
Add support for a CommissioningView with no sections defined, and show
help if neither sections nor the gda.beamline.scripts.procedure.dir
property is set.

Replace underscores in script names with spaces to generate button text.

Add documentation, an example CommissioningViewFactory Bean and an example
plugin.xml element.

Finally, implement the unimplemented changes suggested by the previous
review of the ScriptChooserView, all of which apply to the CommissioningView.

Change-Id: I8d4123d784b8aca3787d3f68eedae84a958b8efc