[DAQ-3818] fix NPE when including configured DummyMonitor in scan
A scan was failing when it included a monitor
which had no-input names, but an extra name. DummyMonitor is an
example of this, once configure() has been called. Normally this would
be called as part of spring initialization. NexusScanDataWriterScanTest
already included DummyMonitor as part of scans, but didn't call
configure() on it.

ScannableNexusDevice did not properly deal with the case where a
Scannable has no input names. In particular the NexusObjectProvider was
not properly configured: the 'primaryDataFieldName' wasn't set. This
causes a problem for scans which include a monitor and no detector,
as the nexus writing framework tries to build an NXdata group using the
primary data field of the monitor. As this wasn't set, an NPE occurred.

Change-Id: Ibae63038c2ea5cb42dbe490697f2bfa079ba04cf