B18 : Minor bugfixes for QEXAFSParametersUIEditor
- Prevent a NPE exception during creation of the QexafsParameters GUI
caused by trying to convert a NaN from energy to wavevector.
This is 'transient' behaviour - the real values are converted later on
and correctly displayed in the GUI.
- Ensure that the 'Element', 'Edge' and 'Edge energy' GUI elements
are updated when the composite is created. This fixes a bug where the
edge energy is not shown when the view is generated.

Change-Id: I39d266f0b4a75436ff4afd171fb6a8da752139d7