I16-508 (9.23) Use plot view ROIs rather than rendering ROIs onto data
ProcessingDetectorWrapper ROIs being rendered onto the dataset of a view
prevented the actual data being seen with a mouse-over and prevented the
automatic histogram functionality working properly.

Using the ROI functionality of the Plot View, where possible, fixes both
of these problems, plus it allows automatic box profiles to be generated
if desired.

ProcessingDetectorWrapper falls back to the old behaviour if no GUI bean
is available, or another error occurs.

Change-Id: I6102cd81098561a3f3ecf68b6e7226b4078b0315