Fix error handling in PositionerRequestHandler
Any problem used to be mangled into an EventException which is really
unhelpful for clients. Now the PositionerRequest has an error message
field which will be empty after a successfully handled request, or
populated with an indication of the problem if unsuccessful.

Addionally, tests have been created for every logic branch, and the
default timeout in the PositionerProxy has been dramatically increased
given that the request is blocking and it could reasonably trigger a
'long' move.

Change-Id: I4d11656e5c4e59b92702942d92ab95d4156927dc