Update metadata handling in NcdStatusModel
Since the switch to RMI, GDAMetadata overrides the source of metadata
updates with the top level metadata object. This means events from the
sample_background were ignored by the status view.

As there is no way to check which metadata entry was updated, this
causes the model to be refreshed for all metadata updates.

See DAQ-3862 and I22-1141

Change-Id: Ia5c87b9f0d9ee930498552a5be69659bc0bab786