[DAQ-3883] fix ClassCastException in NexusScanDataWriter.createScannableNexusDevice
This occurs when there is some object in the jython namespace (not a
Scannable) that has the same name as an INexusDevice in the scan. This
is because we use the same list of names (metadataScannableNames) to
look up both per-scan monitors and pure INexusDevices. For each name in
that list, we look first to see if there is a Scannable with that name
in the jythonnamespace, and if not, we then check to see if there is
an INexusDevice with that name in the INexusDeviceService. Previousy
we cast the object from the jython namespace to Scannable. Now we
do an instanceof test for Scannable. If the test fails, we log a warning
and continue to see if the INexusDeviceService contains an INexusDevice
with that name.

Change-Id: I595f570a619cab562e1068e80f82f5f83c6fa165