DAQ-2832 Ignore the tests as they duplicate others unreliably
Testing of interpretation and submission of ScanRequests and Scan beans
is covered by several other Test classes in
org.eclipse.scanning.test.event and creation via the
ScanRequestBuilderTest. This is an integration test which relies on both
the Jython infrastructure and the scan submission mechanism it invokes
completing their tasks in an alloted time regardless of system load and
network conditions and so is inherently unreliable. Because of this the
test should be ignored for the time being and removed completely along
with the rest of the original mscan implementation which it is aimed at
testing.

Change-Id: I9f17e82ae837ec5ecf12e3a30313cd7da1cf46b0