DAQ-3779 Fix bug with profile loading for certain word order
Because gda_core takes a greedy approach to recognising GDA_COMPONENTs
(by assuming all words in the command not preceded by '-' are
components), when using spring profiles with the profile(s) specified
after e.g. servers, the startup would fail. This is because the profile
name(s) would be interpreted as a component name and not match the
allowed ones causing exit.

In order to use the linux standard of specifying params of an option as
a plain word after the --OPTION text, the parsing logic is changed to
recognise a component name if one is present, ignoring other non
reserved names and only exiting if no component is specified. This
allows the spring profile names to pass through to gda_servers_core to
be interpreted.

This change also updates the help text to remove references to the no
longer supported individual servers invocations and reinstates the
incorrectly removed start/stop/restart interpretation, removing the
PROCSERV_RESTART clause which was the original intention.

Change-Id: Iba14f6494cf292abde3634144033deabf1d91f51