I14-906 Allow loading a 'default' frame when setting up processing chain
GUI has been refactored to have three radio buttons:
- Load default frame: plots the data from the default file displayed
below the button
- Load existing frame: enables Browse button that allows to search for a
file in the directory and displays the path of the file selected.
- Acquire new frame: enables Acquire button that acquires a new frame
with the exposure time, detector and timeout displayed in the GUI.

- Modified update function
When plotting new data when there was data already plotted, the old
dataset would be displayed instead of the new frame selected. By
deleting the Job functions, this problem is now solved and new frames
can be plot.

- Modified loadDataFromFile(String filePath)
There were some errors with the selection of the radio buttons. After
selecting the defaultFileButton that calls loadDataFromFile(String
filePath), other radio buttons had to be selected twice in order to
enable the controls. By removing the the runnable methods in this
function, this has now been solved.

Change-Id: I8f942a83e025473b0d5742602210703ecaefb293