[DAQ-3972] make MetadataFields optionally log error instead of throwing exception
E.g. for ScannableField scannable with the configured name was not found
or there is an error getting the position of that scannable.

Most subclasses will throw an exception by default, and the property
failOnError must be explicitly set to false in order to configure the
field to log an error. ProcessingVariableField (in gda-epics repo) is
the exception, where the default is to log an error.

Change-Id: I6b250204457ee16a58eee2aa932d0492d1452bbc