DAQ-486 Remove wildcard imports in GDAJythonInterpreter initialisation
Imports here are forced to be performed in every GDA instance but
they may not be required.

Since switch to Java 11 in plugin BREE (made for all projects from
GDA 9.21 onwards) these imports would not perform as expected due to
Jython 2.7.1 not supporting Java 11 bytecode.

As this misbehaviour has not caused concern we have some confidence
that they can be removed.

See also DAQ-3079 and DAQ-4053.

Change-Id: Ia2e08f560fb6cce219835faef414efd1468c90ad