Add deletion version for deprecated ScriptProjectType EPICS
This has been deprecated since 2011. Adding this logging should be
enough to convince us that it can finally be deleted.

The standard cycle of deprecating for two releases has not been used
as this type has already been deprecated for so long and there are no
references to it in beamline Spring configuration.

Change-Id: Id4a426b288ab9792012d3942657cd017b5f6f151