Use annotation builder for all services in u.a.d.d.server.configuration
Including all the annotations in the code adds clarity e.g.
@Activate annotations.

Also refactor code to obtain valid configuration service from the list
of available services. Previously this relied on the fact that
components were created in a specific order (due to the MANIFEST.MF).
Now, using a proper reference, all instances of configuration services
are injected.

The ComponentServiceObjects type is useful as
it allows the service to be obtained in addition to the ServiceReference
(used to get properties).

Change-Id: Ib53b7cad8e4fa5de33d8a0c0460b2b6d33b97797