Fix CalibrationFrameCollector
This child class implements configureCollection() in order to create a
fresh FrameRequestDocument with every scan. An early return (needed for
K11) was erroneously inserted which caused scans to run with potentially
outdated FrameRequestDocuments in other beamlines. This is now fixed,
with most of the logic pushed up to the parent class. K11 config will
now instantiate the parent class - other beamlines can use the child
class.


Change-Id: I880c4dbdb5d1ec261e13b090b60318ef4054a0ab