Combine configure with constructor in GDAJythonInterpreter
No longer call initialise on construction/configure becuase in some
contexts, e.g. unit tests we don't want to instantiate a new translator
and import things into the namespace.

Also initialise the translator with a no-op translator in case something
later tries to translate something.

The configure method was always required anyway to actually use the
interpreter. Removing this also avoids confusion with the configure
method on the Configurable interface which this class does not
implement.

Future changes may split the constructor into smaller methods and
possibly move the builtins/GdaGlobals to initialise.

Change-Id: I855f041215b38c895bc3f0f7eba5c30e49d7700e