DAQ-3079 Add back bundle iteration in GDAJythonInterpreter
Ask the OSGi framework for the present bundles in the application
and get their location. Which allows for (almost) the same handling
regardless of whether in the IDE or not.

Note that wildcard imports and package discovery is re-enabled with this
change which was made possible by the new version of Jython which is
able to read Java 11 bytecode.

Change-Id: Ic48363fbaddf8d6a7a4ce46655f5e761b8cd0071