I22-1157 Rearrange shutter controls on NCD views
Rearrange the layout of views to improve use of space and allow the
order of shutters to be controlled.

To set the layout, a shutterOrder property should be set to a comma
separated list of colon separated lists. Each element of the outer list
is a row on the view and each element of the inner lists are enum
positioners on the row.

The composite containing the shutter controls is also extracted to its
own class to reduce the duplication between the NcdButtonPanelView and
NcdStatus views.

Change-Id: I5adb13deee26adab89ab659e07eb898a61d7a65a