improve generic for UnitsProvider by replacing wildcards with type variable
also use Set.of instead of google guava's ImmutableSet.of

Note that the method getCompatibleUnits has been refactored to use
a return variable of Set<?> instead of multiple-return statements.
This is because each return statement would need two casts, one
to Object (not (Set<?>) doesn't even work there) then to Set<Unit<?>>.
It is not possible in Java to assign Set<Unit<Length>> to a Set<Unit<?>>
variable.

Change-Id: If18ea360c7467a9504a9f6d64a34722c353dd866