I15_1-1177 (9.25) Add stop support to GDAJythonScriptService
Use runsource() in GDAJythonScriptService.execute() so that scripts are
interrupted when JythonServer.abortCommands() called.

Using evaluateCommand() means that Solstice scans cannot be stopped if
they run a before or after scripts which doesn't return for any reason.
When this happens, the whole GDA servers has to be restarted in order to
clear the blocked scan.

Using runsource() means that the thread running the the script is
tracked by JythonServer. Note that runScript() and runCommand() are not
appropriate as they do not block.

Change-Id: Iad463ba0716b19a1fe0c91d7c2a7db1e4540d48f