I15-685 I15-684 (9.25) Fix LiveControlsView.reopenWithSecondaryID
If a name for a LiveControlsView had been set in a plugin.xml, it would
only display before the view was activated for the first time. After
activation it would revert to the name of the controlSet which
configured it.

Since we know the views Registered name before reopenWithSecondaryId is
called, we can record it, and restore it when the new view is created.

Change-Id: I66a3a0477fac8c20c273b8b45782306364a8001d