DAQ-4071 : Improvements to make XSpress4 PV names configurable
- A list of all PVs used by the Xspress4 controller are now stored in a
new Enum (XspressPvName). This contains the default name/template to be used
when creating the PV objects.
- The PV names used in the Epics controller are looked up from a map from
XspressPvName to name/template string. The values in the map are initialised
from the XspressPvName enum. If required they can be also be customised by
passing a map of replacement values to the controller if required.

Change-Id: I36f0c260eeec4b68e4d9559ef9ae0b554f6b5099