I14 Add update points call when selecting a xanes edge element
- Fixes bug:

When switching from mapping to xanes scan type, the checkbox is
preselected but the scan path editor is empty.

Then, when selecting a xanes edge, the scan path editor gets populated,
but there is no event in the checkbox binding that calls to revalidate
the model and recalculate the scan path.

- This adds:

Call to revalidate the model and recalculate the scan path when a xanes
edge is selected.

Change-Id: Ib482956784e48118c371c002ae4a4574c0cf78e4