DAQ-3993 Allow bean def of NexusDataWriterConfiguration to work
The idea was to define NexusDataWriterConfiguration in Spring but until
now this has not properly worked due to the possibility of multiple
instances of NexusDataWriterConfiguration being created.

Now we enforce that only one instance is made (via checking/registering
a service with OSGi). This implies that beamline Spring must use either
NDW static methods (with MethodInvokingFactoryBean) or a bean definition
for NexusDataWriterConfiguration.

Change-Id: Icb8590023e1ff226a953ee59bbadd2b5a66f2ea7