Reduce log level for CNFE in RmiAutomatedExporter
If the class cannot be loaded (class not in application, or mistake in
FQCN) the method copes by returning an empty set and calling code may
proceed based on this.

In addition drop the exception from the message as this method is only
called over RMI so is largely unhelpful.

Change-Id: I620eb48ed17af85badd7ec90b895f64827346a98