K11-1563 Rename DetectorDocument method and set consistently
DetectorDocument.name was previously set both from DocumentFactory and
ScanningAcquisitionDetectorUpdater as the name of the CameraControl bean
which caused some confusion since the method implied it would return a
detector name.

Now we set it as CameraConfigurationProperties.getId(), and the method
name is getId() instead of getName(), which is more consistent with its
current usage.

Change-Id: Ic4e65d14559cecdb04e8a30760deb70ab664f139