DAQ-4098 Use bean ID in gdaserver.py module
Changes to the FindableNameSetterPostProcessor mean that the name of a
findable is no longer required to be the same as the ID it is found by
in the finder. When the Finder writes its gdaserver module it should use
the identifier that the object can be found by rather than the name
returned by getName().

Change-Id: I03e61beef9fd4c519e1177d1f2f722183f8425a7