I14 Removing check for suspended queue in a label provider
- I14 saw errors from gda saying "Could not get consumer info". This is
due to the recent change where the LabelProvider for the Location column
checks whether the queue is suspended or not. It timesout and Eclipse
display this message.

- UpdateQueueJob does not get called enough to cache the boolean value
of whether the queue is paused.

- The queue view will still gray out when the queue is suspended but the
Location column won't have a gray colour and will still be black or blue
whether the scan was finished.

Change-Id: I49b60a9e802fe0f9d254d0813895dc1c119e6979