check for non-point-dependent nodes in writeDataNode
previously this writeDataNode was being called for every node in the
tree, even non-point-dependent nodes (i.e. metadata nodes).
This still worked as that method checks that a dataset has been created
for the node, which is not the case for a non-point-dependent node
and if not it logged an error. This was due to the assumption that this
could only happen if a subsequent tree contained a node not present in
the first tree for this first point. This assumption was wrong.

Change-Id: I88fc672ab516d4448b6ca61dab20877eedcd635d