DAQ-4108 Restore extended array syntax for TokenStreamTranslator
The GeneralTranslator supported using [lists without commas] in
commands. This was originally omitted from the token based
implementation as it introduces several ambiguities but as several
beamlines had come to rely on it, it has been reimplemented here as
well. Arrays without commas are expanded to Jython lists only in cases
that are already being translated (as part of an aliased command) to
reduce the risk of introducing errors to valid Jython syntax.

eg in vanilla Jython `[8 -3]` is equivalent to `[3]` whereas in
translated commands it should be interpreted as `[8, -3]`.

Complex constructs are not supported in the extended list syntax and
should be wrapped in () if they are required, eg (if foo is an alias)

    foo [(lambda x, y: x+y) ([i for i in range(10) if i % 3])]

Change-Id: I9996d386cf684183823e13df62ef6dc7fa5060bb