add IAxialModel interface for single axis scan path models
This can be used in the mapping bean and other parts of the code to
limit the path models for outer scannables to models for a single
axis.

A new abstract class, AbstractAxialStepModel was created, which
AxialCollatedStepModel and AxialStepModel now both implement, with
AxialStepModel adding no additional functionality, but which implements
IAxialModel, whereas AxialCollatedStepModel does not. This is because
AxialCollatedStepModel represents multiple axes moving together, which
is not appropriate for IAxialModel.

Change-Id: I214389e3ed37dfe560a4044ee95fd72f8046ecc5