K11-1579 Fix start positions bug
Start positions (defined for all acquisitions, per acquisition type, per
acquisition instance) were getting mixed up. This is because the initial
set was mutated in the first method, and the mutated instance given to
the next method, where it clearly wanted the same starting point and the
first. This is fixed by both fetching the initial set of positions in
the body of each method, and also wrapping in a mutable HashSet since we
cannot know the mutability of the original set.

Change-Id: I67618c73225b5c5dee2c3913fe70caecfc8f69dd