DAQ-4119 Rename BatonManager getAuthorisationLevelOf
The value returned was dependent on other factors (whether the baton was
held and whether RBAC was enabled) and could return different values to
method of the same name that takes a client index instead of name. It
was also different to the similarly named getAuthorisationLevel method
of ClientInfo.

This renames the misleading method to effectiveAuthorisationLevelOf and
removes the almost entirely unused index method.

Change-Id: I1746cda1416c79ba1fb83165f36ad78d604ca7c0