B16-344 Ensure all images for IPP scan are in same directory
The IPP plugin has a limited api which forces a directory to be created
for the current date. This was previously called for every acquisition
but this led to images for a single scan being split across two
directories if the scan spanned midnight.

Instead only call this at the start of the scan.

Change-Id: I3292adbf397f19ff9f9647f7e2e2fe032eafa7c3