K11-1602 Add Payload, service and handler, and ScanningAcquisition impl
This change introduces the classes that will replace Triggerable logic
in an upcoming change.

Payload, unlike Triggerable, is a purely marker interface and represents
some data or request. The handling of this is performed by a particular
PayloadHandler, registered with the Payload type in the PayloadService.
The handler can also wrap the request into its Payload form.

This change implements the above classes for ScanningAcquisition. In the
upcoming change, this would replace TriggerableMap.

Change-Id: If08fcabb5d4447160caa65760ecd7045b5de6ff2