I15_1-1148 (9.25) Fix mismatch between button status, icon and tooltip
The icon and tooltip were only being updated when the button was pushed
in the Gui. Update them when status is updated from elsewhere too.

Also, change the icons to be more visually distinct and move the button
to the start of the toolbar, where it is more prominent.

Since this icon was acting as a separator between submitted/running and
completed jobs, add a separator where it was.

Fix a typo in the downAction tooltip.

Make action update code more consistent.
Change-Id: I40b63a61c3cde69b4c4481630bc647aeff4a7a46

Change-Id: I5af6067f2edd7a423052d32f612a95c1a130f953