DAQ-4116 Summarise a map scan quickly, when possible
StatusPanel displays the estimated number of map points for a rectangle
with a grid while it waits for the exact number of points from PathInfo.
Then, the label is rewritten.

- RegionAndPathController creates a PathInfoEstimator object with the
scanPathModel and the scanRegionShape and sends this as an event to the
MappingExperimentView.

- PathInfoEstimator checks the whether the scanPathModel is a
TwoAxisGridPointsModel and the scanRegionShape is a Centred Rectangle to
get the number of points for x and y axis.

- MappingExperimentView sets this PathInfoEstimator in the StatusPanel.

- StatusPanel gets the estimated 2d points from PathInfoEstimator and
multiplies this number by the Outer Scannables number of points. It also
displays the exposure time multiplied by the total number of points.


Change-Id: Ic11ca24cfec7edc2698f4c2211ebddb116c56c53