I15_1-1177 (9.25) Improve ICommandRunner documentation
In particular, make it clear whether evaluateCommand, runCommand,
runScript or runsource are Blocking, Interruptible and/or script locked
and explain what effects these options have.

Use of the wrong call can result in commands and scripts ignoring
interrupts, blocking or running asynchronously when they shouldn't,
or running when another script expects to be running exclusively.

Change-Id: Ie57944da7804fdf2627572496f4b48d25c20504c