Add IObservable updates support to PositionerDeterminedScannable
This change first automates the switching of active scannable in res-
ponse to selector moves.

Secondly, an observer is attached to the active scannable (and removed
from previous active scannable) to forward the inner scannable's events
to PositionerDeterminedScannable's observers.

Finally, we also notify observers when the active scannable changes,
which is is probably a significant event for observers.

Change-Id: I350b173d03df3ff4794e66db5c09b5fa86698a84