I08-396 Add section to submit energy scans as multiple 2d scans (I08-1)
- Creates new section in the Mapping GUI of I08-1 that appears when
the option 'Energy scan' is selected on the scan type list of 'Select
Scan Type'.

- An element/edge dropdown box is available to select one element and
populate the outer scannable energyFocus.

- When the button 'Run energy scan' is selected, the function
submitScan() gets the energy model from the energyFocus selected
parameteres. Then, only the 2D scan is set as the compound model in
the scanRequest object.

- Both scanRequest and energyModel are serialized as JSON objects into
the Jython namespace as scanRequestJson and energyScanParamsJson



Change-Id: I2b59056aa4d78d0e14629e8b7b44e179c2f48adc