K11-1641 Let ExperimentController use standard naming for scan files
NexusExperimentController now generates standard file paths for scan
files. It still produces experiment node files under @visit/experiments
using a directory hierarchy. Leaf nodes take the name from the acquisi-
tion, but link to nexus files saved as @visit/nexus/k11-####.nxs.

To achieve this, NodeFileCreationRequest now has a Map<String, URL> to
represent a node's children (instead of simply Set<URL>), allowing the
name-absent in the URL-to be set in the external link.

To simplify file paths in general, a NumTracker is used instead of time-
stamps to ensure unique node files. Consequently the check for duplicate
experiment files can be removed (this was due to the timestamp format
which only repesented the date meaning two experiments with the same
name started on the same date resulted in equal file paths...).

Tests modified to reflect this change, and logic in NexusExperiment-
Controller simplified where possible.

Change-Id: I9994fd67e845f04329ef3a8cb0f1b3d7daa860a9