Add Java exception handling to pytest runner
If a pytest test case raised a Java exception, it was not being caught
by the normal mechanism and the test was being marked as having passed.

This adds a pytest plugin that wraps each test case in a try/catch block
to fail on any Java Throwable.

See I15_1-1175

Change-Id: I7de8798dadcfaaead0aafaa9342b18b32fe7caac