I14 Add energy offset to dcm_enrg values
When user selects an element/edge, the dcm_enrg values gets populated
with a range of start, stop and step size models around the energy at
that edge.

This text box allows to enter an energy offset (eV) that will move the
energy of each element slightly, and then the ranges will be generated
from that value instead.

Change-Id: I6775351cacfb2f517fc0e6a8f0802a3eeb88b5b2