fix toString for AbstractMultiStepModel and AbstractPointsModel
for AbstractMultiStepModel the toString method in the superclass already
prints the class name and the child models. The toString method in
AbstractMultiStepModel calls this superclass method, duplicating the
child models in the string. This change removes this method, since the
superclass method already does what we want.

The toString method in AxialPointsModel is fixed to call the superclass
toString method to show common properties

Change-Id: I64b5f56c7cbd8e5408fedd39f1af59d5cdfd7d4a