Save AcquisitionKeys in ScanningAcquisition
AcquisitionKeys are originally defined by the various scan UIs and
loaded into the AcquisitionController alongside, but separate to, the
Acquisition. Until now only the acquisition gets saved; when a previous
acquisition is loaded, the keys are guessed by evaluating acquisition
fields, but as the acquisitions become more complex this step often
fails.

The simple solution is to save the AcquisitionKeys in the acquisition
document itself. This requires moving this and related classes from UI
projects to uk.ac.gda.api.acquisition.

Change-Id: Ibe24b697922fa105961f399406ccae028624b236