Throw exception if NumTracker can't increment scan number
If something goes wrong while incrementing the scan number, this should
be propagated to the caller rather than returning a default value.

Returning 0 when the file can't be written means that the error seen by
the user is that a file with scan number 0 can't be overwritten which
can be misleading when it's not the scan number that should be being
used.

Change-Id: I7f724440128a431e302859ad597edc05bcb10f20