I22-505 Initial implementation of Bragg Calibration tools
Consists of a server side that positions the edges and runs scans and a
client component that allows users to select which edges to scan, and
then compare the position of expected features in the scan data against
the observed positions. These values are plotted against each other and
a linear fit is used to determine the required offset for the bragg
motor.

While this is expected to require changes after use on the beamline,
this version is intended as fulfulling the minimum requirements of the
calibration process.

Change-Id: Ib3ee70104669add1c4bd74852f88f6f9f0847d18