I22-1211 Improve tetramm valuesPerReading calculation
Instead of imposing a fixed number of samples per frame, set a maximum
number of samples per frame. This allows the frame size to be reduced
automatically without having to reduce the nummber of readings to below
the minimum requrired to even out noise in the data.

It was previously possible to make these changes manually but changing
the framesize automatically means the framesize is always as close to
the target as it can be.

Change-Id: I059d65ab950901f4b83abd5bfbdc095930b9f090