Support pytest marks in pytestrunner
When tests are decorated with @pytest.mark.NAME, they can be skipped or
included based on the mark name.  To include all tests marked `foo`,
pass 'foo' to `run_tests` using the `marks` keyword.

    >>> run_tests(marks='foo')

To exclude all tests marked with `foo`, the marks can be negated, eg

    >>> run_tests(marks='not foo')

Marks can be combined using 'and' and 'or' as required

    >>> # run tests marked foo unless they're also marked bar
    >>> run_tests(marks='foo and not bar')
    >>>
    >>> # run all tests marked either foo or bar
    >>> run_tests(marks='foo or bar')
    >>>
    >>> # only run foo tests if also bar
    >>> run_tests(marks='(foo and bar) or not foo')

Change-Id: I4413649687c7ee3383ff4103faa2d17b75b30cf5