K11: Extract scan position management into own class
The overall start position of a scan in K11 is created from several
sources of configuration. The methods dealing with it were unclear,
and the beamline has struggled with position-related bugs for a while.

In order to simplify the logic, these methods are extracted into a new
unit-tested class PositionManager.

Change-Id: Iba421d83dbcbafe54f1afcc898666bf81cb044a3