[DAQ-4259] add handling for jython types when getting values to write for scannable fields
AbstractScannableNexusDevice and BeforeScanSnapshotWriter were not
previously able to write fields whose value was a jython object such as
a PyFloat.

ScannableUtils has an existing method to deal with jython
types, objectToArray, but that method tries to convert all values to a
double - it returns a Double[]. This means we can't use it as we need
to allow String valued scannables.

A new method ScannableUtils.toObjectArray was created that deals with
python objects, but also preserves Strings. This is now used by
AbstractScannableNexusDevice and BeforeScanSnapshotWriter to get the
value to write to nexus for each field of the scannable.

Change-Id: I2c72f3c54ad200dcd1294b7e89ea89cb5f9d1e22