I07-475 Prevent flickering in JythonTerminalView
Instead of jumping to the end of the text and then scrolling to maintain
horizontal position, set the final line as the top line in the view. As
it's not possible to scroll that far, this leaves the view scrolled to
the end of the output without affecting the sideways scrolling.

As the setTopIndex sets the position of the top line, the final line can
be partially obscured if the view is not an exact multiple of the line
height. To prevent this, a control listener is added to adjust the
bottom margin to keep the usable height to an exact number of lines.

Change-Id: Ieddb47d1de92ef9de54865283ff114421db99f76